Move RtcApi to embassy-stm32/src/rtc/mod.rs
