--- examples/common/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-edition = "2024"
-name = "embassy-examples-common"
-version = "0.1.0"
-license = "MIT OR Apache-2.0"
-publish = false
-
-[dependencies]
-defmt = { version = "1.0.1", optional = true }
-
-[features]
-default = []
-defmt = ["dep:defmt"]
-
