--- examples/common/src/lib.rs
@@ -1,3 +1,0 @@
-#![cfg_attr(not(test), no_std)]
-
-pub mod rtc_api;
