+++ examples/stm32wba6/Cargo.toml
@@ -12,7 +12,6 @@ embassy-executor = { version = "0.9.0", path = "../../embassy-executor", feature
 embassy-time = { version = "0.5.0", path = "../../embassy-time", features = ["defmt", "defmt-timestamp-uptime", "tick-hz-32_768"] }
 embassy-usb = { version = "0.5.1", path = "../../embassy-usb", features = ["defmt"] }
 embassy-futures = { version = "0.1.2", path = "../../embassy-futures" }
-embassy-examples-common = { path = "../common", features = ["defmt"] }
 
 defmt = "1.0.1"
 defmt-rtt = "1.0.0"
@@ -26,7 +25,6 @@ static_cell = "2"
 
 [dev-dependencies]
 embassy-stm32 = { version = "0.4.0", path = "../../embassy-stm32", features = ["stm32wba65ri"] }
-embassy-examples-common = { path = "../common" }
 
 [profile.release]
 debug = 2
